refactor: GearChainPoleConnectを上・中・下の3層ディレクトリに整理
- ルート: GearChainPoleConnectSystem（エントリ、モード分岐のみ）
- Modes/: アイテム別モードロジックと共有Context
- Parts/: モードから使われる部品（判定・計算・表示・通信・コライダー）
- アイテム種別判定をGearChainPoleItemFinderとしてPreviewCalculatorから分離
  （PlaceSystemSelectorのシステム選択がプレビュー計算クラスに依存していた混線を解消）
