+++ moorestech_client/Assets/Scripts/Client.Game/InGame/BlockSystem/PlaceSystem/GearChainPoleConnect/Parts.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0366fc5e5719c47668cb01a3d65fb323
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
