+++ moorestech_client/Assets/Scripts/Client.Game/InGame/BlockSystem/PlaceSystem/GearChainPoleConnect/Parts/GearChainPoleItemFinder.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: a6b18e74a9d5144c799f166c20318bf6
